Add files via upload
BMI Calculator

This repository contains a simple and user-friendly Body Mass Index (BMI) calculator. The program calculates the BMI based on the user's weight (in kilograms) and height (in feet), and categorizes the result into underweight, normal weight, overweight, or obese. It is a handy tool for anyone looking to monitor their health and fitness.